+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-06 11:03:51
-// Commit: chore(core): update dependencies and cleanup unused imports in core
+// Updated at: 2026-04-06 11:44:47
+// Commit: refactor(core): optimize data flow between domain and data layers
 
 class coreComponentState {
-  final String commitRef = '150';
-  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
-  final DateTime timestamp = DateTime.parse('2026-04-06T11:03:51+05:00');
+  final String commitRef = '152';
+  final String lastMessage = 'refactor(core): optimize data flow between domain and data layers';
+  final DateTime timestamp = DateTime.parse('2026-04-06T11:44:47+05:00');
 }